rewards rework in the works, attempting to locate sources for proper display
