+++ controllers/profileRoutes.js
@@ -1,8 +1,20 @@
+const { User } = require('../models');
+
 const router = require('express').Router();
 
 router.get('/', async (req, res) => {
-  // Send the rendered Handlebars.js template back as the response
-  res.render('profile');
+  try {
+    console.log("Does it work?")
+    const dbUserData = await User.findOne({ where: { id: req.session.userId} });
+    console.log(dbUserData)
+    console.log("Does it work?")
+    res.render('profile')
+    console.log("Does it work?")
+    
+  } catch (err) {
+    console.log(err);
+    res.status(500).json(err);
+  }
 });
 
 module.exports = router;
