+++ views/checkout.handlebars
@@ -1,5 +1,5 @@
 <div class="row">
-        <form class="col s12">
+        <form class="col s12" id="checkoutForm">
 
             <div class="row">
                 <div class="input-field col l6 offset-l3 m8 offset-m2 s10 offset-s1">
@@ -125,7 +125,8 @@
           
           <div class="row" id='marginBottomSmall'>
             <div class="center-align col s8 offset-s2">
-                <a class="waves-effect waves-light btn-large" id="submit">Checkout</a>
+                <button class="waves-effect waves-light btn-large" type="submit" name="action">Checkout</button>
+                <input id="payment"></input>
             </div>
         </div>
 
@@ -134,3 +135,4 @@
 
 
     <script src="./js/checkout.js"></script>
+    <script src="./js/rewards.js"></script>
