+++ views/profile.handlebars
@@ -1,5 +1,5 @@
 <div class="row container" id="personalInfo">
-        <form class="col s12">
+        <form class="col s12" id="checkoutForm">
 
             <div class="row">
                 <div class="input-field col l6 offset-l3 m8 offset-m2 s10 offset-s1">
@@ -112,3 +112,4 @@
 
 
     <script src="./js/profile.js"></script>
+    <script src="./js/rewards.js"></script>
